Fixed obj reader and remove a branch at triangle hit

+++ Reza_Raytracer/OBJReader.h
@@ -20,7 +20,7 @@ public:
 	void ReadOBJ(const std::string& filename,
 				 std::vector<Vec3>& vertices,
 				 std::vector<Vec3>& normals,
-				 std::vector< std::vector<int>>& uvs,
+				 std::vector< std::vector<double>>& uvs,
 				 std::vector< std::vector<int>>& vertex_indices,
 				 std::vector< std::vector<int>>& normal_indices,
 				 std::vector< std::vector<int>>& uv_indices);
